Refactor promo code deletion logic in StoreListView
Added a check to ensure `PromoCodeModel` exists before iterating its objects to delete expired promo codes. Also removed unnecessary blank lines in `settings.py` for cleaner formatting.

+++ store/views.py
@@ -140,9 +140,10 @@ class StoreListView(ListView):
     paginate_by = 9
 
     def get(self, request, *args, **kwargs):
-        for promo in PromoCodeModel.objects.all():
-            if promo.promo_end_date < now():
-                promo.delete()
+        if PromoCodeModel:
+            for promo in PromoCodeModel.objects.all():
+                if promo.promo_end_date < now():
+                    promo.delete()
         return super(StoreListView, self).get(request, *args, **kwargs)
 
     def get_queryset(self):
